+++ NEWS.md
@@ -1,5 +1,6 @@
 # shinyalert 1.0.0.*
 
+- New feature (#12): support chaining modals, you can now call a shinyalert modal in the callback of another modal
 - New feature (#11): `inputId` parameter added, which allows you to set the input ID that is used to retrieve the return value of the modal
 - Slight documentation changes
 - small change to demo shiny app UI